# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TournaX Gaming Tournament API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/explore:
    get:
      operationId: exploreUsers
      tags: [users]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Explore hosts and players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploreResponse"
  /users/{handle}:
    get:
      operationId: getUserProfile
      tags: [users]
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{handle}/follow:
    post:
      operationId: followUser
      tags: [users]
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{handle}/unfollow:
    post:
      operationId: unfollowUser
      tags: [users]
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/me/squad:
    get:
      operationId: getMySquad
      tags: [users]
      responses:
        "200":
          description: My squad members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SquadMember"
    post:
      operationId: addSquadMember
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSquadMemberRequest"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SquadMember"
  /users/me/profile:
    put:
      operationId: updateMyProfile
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /wallet/add-balance:
    post:
      operationId: requestAddBalance
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBalanceRequest"
      responses:
        "200":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
      responses:
        "200":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /messages:
    get:
      operationId: getMessages
      tags: [users]
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /complaints:
    post:
      operationId: raiseComplaint
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintRequest"
      responses:
        "200":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        handle:
          type: string
        avatar:
          type: string
        game:
          type: string
        gameUid:
          type: string
        role:
          type: string
          enum: [player, host, admin]
        balance:
          type: number
        status:
          type: string
          enum: [pending, active, banned]
        followersCount:
          type: integer
        followingCount:
          type: integer
        profileSetup:
          type: boolean
      required: [id, email, role, balance, status, profileSetup]
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        handle:
          type: string
        avatar:
          type: string
    Match:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        game:
          type: string
        mode:
          type: string
        teamSize:
          type: integer
        entryFee:
          type: number
        prizePool:
          type: number
        startTime:
          type: string
        status:
          type: string
          enum: [upcoming, live, completed]
        slots:
          type: integer
        filledSlots:
          type: integer
        hostHandle:
          type: string
        hostName:
          type: string
        hostAvatar:
          type: string
        hostFollowers:
          type: integer
        isFollowingHost:
          type: boolean
        isJoined:
          type: boolean
        roomId:
          type: string
        roomPassword:
          type: string
        roomReleased:
          type: boolean
        hostId:
          type: integer
      required: [id, code, game, mode, teamSize, entryFee, prizePool, startTime, status, slots, filledSlots, hostId]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        handle:
          type: string
        avatar:
          type: string
        role:
          type: string
        followersCount:
          type: integer
        followingCount:
          type: integer
        rating:
          type: number
        matchesCount:
          type: integer
        isFollowing:
          type: boolean
        upcomingMatches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
        activeMatches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
      required: [id, handle, followersCount, followingCount, matchesCount, isFollowing, upcomingMatches, activeMatches]
    ExploreResponse:
      type: object
      properties:
        recommendedHosts:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        mostActivePlayers:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
      required: [recommendedHosts, mostActivePlayers]
    SquadMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        uid:
          type: string
      required: [id, name, uid]
    AddSquadMemberRequest:
      type: object
      properties:
        name:
          type: string
        uid:
          type: string
      required: [name, uid]
...
    AddBalanceRequest:
      type: object
      properties:
        utrNumber:
          type: string
        amount:
          type: number
        receiptUrl:
          type: string
      required: [utrNumber, amount]
    WithdrawRequest:
      type: object
      properties:
        amount:
          type: number
        upiId:
          type: string
      required: [amount, upiId]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        fromHandle:
          type: string
        fromName:
          type: string
        content:
          type: string
        createdAt:
          type: string
        read:
          type: boolean
      required: [id, fromHandle, content, createdAt, read]
    SendMessageRequest:
      type: object
      properties:
        toHandle:
          type: string
        content:
          type: string
      required: [toHandle, content]
    ComplaintRequest:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
      required: [subject, description]
...