# lib/api-zod/src/generated/types/exploreResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TournaX Gaming Tournament API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface ExploreResponse {
  recommendedHosts: UserProfile[];
  mostActivePlayers: UserProfile[];
}